+++ lightly_studio/src/lightly_studio/api/db_tables.py
@@ -22,6 +22,9 @@ from lightly_studio.models.dataset import (
 from lightly_studio.models.embedding_model import (
     EmbeddingModelTable,  # noqa: F401, required for SQLModel to work properly
 )
+from lightly_studio.models.embedding_model_service import (
+    EmbeddingModelServiceTable,  # noqa: F401, required for SQLModel to work properly
+)
 from lightly_studio.models.image import (
     ImageTable,  # noqa: F401, required for SQLModel to work properly
 )